Continuity test performed and doc updated.

+++ Slotted_Random_Access_Wireless_Network/src/Transmitter.cc
@@ -166,7 +166,7 @@ void Transmitter::handleChannelPacket(cMessage* msg){
 }
 
 void Transmitter::computeModuleStatistics(){
-    emit(numPacketOnBufferSignal_, buffer.size());
+    emit(numPacketOnBufferSignal_, (long)buffer.size());
 }
 
 void Transmitter::finish(){
